update test utility around TestInvokeMethod
