+++ tests/aws/services/apigateway/conftest.py
@@ -1,3 +1,5 @@
+from itertools import count
+
 import pytest
 from botocore.config import Config
 
@@ -238,3 +240,60 @@ def import_apigw(aws_client, aws_client_factory):
 def apigw_add_transformers(snapshot):
     snapshot.add_transformer(snapshot.transform.jsonpath("$..items..id", "id"))
     snapshot.add_transformer(snapshot.transform.key_value("deploymentId"))
+
+
+@pytest.fixture
+def apigw_test_invoke_response_formatter(snapshot):
+    snapshot.add_transformers_list(
+        [
+            snapshot.transform.key_value(
+                "latency", value_replacement="<latency>", reference_replacement=False
+            ),
+            snapshot.transform.jsonpath(
+                "$..headers.X-Amzn-Trace-Id", value_replacement="x-amz-trace-id"
+            ),
+            snapshot.transform.regex(
+                r"URI: https:\/\/.*?\/2015-03-31", "URI: https://<endpoint_url>/2015-03-31"
+            ),
+            snapshot.transform.regex(
+                r"Integration latency: \d*? ms", "Integration latency: <latency> ms"
+            ),
+            snapshot.transform.regex(
+                r"Date=[a-zA-Z]{3},\s\d{2}\s[a-zA-Z]{3}\s\d{4}\s\d{2}:\d{2}:\d{2}\sGMT",
+                "Date=Day, dd MMM yyyy hh:mm:ss GMT",
+            ),
+            snapshot.transform.regex(
+                r"x-amzn-RequestId=[a-f0-9-]{36}", "x-amzn-RequestId=<request-id-lambda>"
+            ),
+            snapshot.transform.regex(
+                r"[a-zA-Z]{3}\s[a-zA-Z]{3}\s\d{2}\s\d{2}:\d{2}:\d{2}\sUTC\s\d{4} :",
+                "DDD MMM dd hh:mm:ss UTC yyyy :",
+            ),
+            snapshot.transform.regex(
+                r"Authorization=.*?,", "Authorization=<authorization-header>,"
+            ),
+            snapshot.transform.regex(
+                r"X-Amz-Security-Token=.*?\s\[", "X-Amz-Security-Token=<token> ["
+            ),
+            snapshot.transform.regex(r"\d{8}T\d{6}Z", "<date>"),
+        ]
+    )
+
+    def _transform_log(_log: str) -> dict[str, str]:
+        return {f"line{index:02d}": line for index, line in enumerate(_log.split("\n"))}
+
+    counter = count(start=1)
+
+    # we want to do very precise matching on the log, and splitting on new lines will help in case the snapshot
+    # fails
+    # the snapshot library does not allow to ignore an array index as the last node, so we need to put it in a dict
+
+    def transform_response(response: dict) -> dict:
+        response["log"] = _transform_log(response["log"])
+        request_id = response["log"]["line00"].split(" ")[-1]
+        snapshot.add_transformer(
+            snapshot.transform.regex(request_id, f"<request-id-{next(counter)}>"), priority=-1
+        )
+        return response
+
+    return transform_response
